A working use of prism to highlight code by depth
Only used on lisp.

It replaces rainbow-delimiters-mode for supported modes.
There is a list of similar modes mwb-lisp-modes and mwb-braces-modes
but lipy load does not use that as too complex.

lip.org altered so it compiles - smartparens now has hydra as part of
use-package.

This required adding use-package-hydra as I could nopt be bothered to
convert to pretty-hydra.